Support textures on the web as well
