+++ widgetry/src/shaders/fragment_300.glsl
@@ -7,10 +7,14 @@ precision mediump sampler2DArray;
 uniform vec3 transform;
 // (window width, window height, z value)
 uniform vec3 window;
+// textures grid
+uniform sampler2DArray textures;
 
-in vec4 pass_style;
-out vec4 f_color;
+in vec4 fs_color;
+in vec3 fs_texture_coord;
+
+out vec4 out_color;
 
 void main() {
-    f_color = pass_style;
+    out_color = fs_color * texture(textures, fs_texture_coord);
 }
